requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.27.0
pydantic==2.4.2
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
pytz==2024.1
pytest==7.4.3
pytest-asyncio==0.23.2
httpx[http2]==0.27.0
